add computed properties with cache invalidation for reactive UI updates

+++ app/Livewire/ArticleIndex.php
@@ -5,6 +5,7 @@ namespace App\Livewire;
 use App\Models\Article;
 use Livewire\Component;
 use Livewire\WithPagination;
+use Livewire\Attributes\Computed;
 
 class ArticleIndex extends Component
 {
@@ -12,9 +13,27 @@ class ArticleIndex extends Component
 
     public $showOnlyPublished = false;
 
+    #[Computed]
+    public function articles()
+    {
+        $query = Article::query();
+
+        if ($this->showOnlyPublished) {
+            $query->where('published', 1);
+        }
+
+        return $query->paginate(8, pageName: 'articulos-pagina');
+    }
+
     public function delete(Article $article)
     {
+        if ($this->articles->count() < 8) {
+            throw new \Exception("No");
+        }
+
         $article->delete();
+        unset($this->articles);
+        cache()->forget('published-count');
     }
 
     public function showAll()
@@ -27,18 +46,5 @@ class ArticleIndex extends Component
     {
         $this->showOnlyPublished = true;
         $this->resetPage(pageName: 'articulos-pagina');
-    }
-
-    public function render()
-    {
-        $query = Article::query();
-
-        if ($this->showOnlyPublished) {
-            $query->where('published', 1);
-        }
-
-        return view('livewire.article-index', [
-            'articles' => $query->paginate(8, pageName: 'articulos-pagina'),
-        ]);
     }
 }
